% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeLRT.R
\name{TreeLRT}
\alias{TreeLRT}
\title{Likelihood ratio test for testing equality means against tree-ordered alternatives in one-way ANOVA}
\usage{
TreeLRT(sample_data, significance_level)
}
\arguments{
\item{sample_data}{list}

\item{significance_level}{numeric}
}
\value{
Critical value numeric

Test statistic value numeric

Result Character
}
\description{
Likelihood ratio test for testing equality means against tree-ordered alternatives in one-way ANOVA
}
\details{
This package gives the likelihood ratio test for assessing the equality of means versus tree-ordered alternatives in one-way ANOVA. Here we considered testing of hypothesis testing of H_0:mu_0 = mu_1 = ... = mu_k vs H_1:mu_0 <= mu_i (with at least one strict inequality), for all i greater than or equal to 1, where mu_0 and mu_i represent the population mean of the control and i-th treatment, respectively. The input comprises two variables: sample_data and significance_level. The output includes the critical value, the LRT statistic value, and the outcome, which shows whether to accept or reject the null hypothesis.  If the value of the test statistic less than the critical value then we reject the null hypothesis; otherwise, we do not reject the null hypothesis.
}
\author{
Subha Halder
}
